Module Httpaf.Reqd

Request Descriptor

type t

val request : t -> Request.t

val request_body : t -> [ `read ] Body.t

val response : t -> Response.t option

val response_exn : t -> Response.t

Responding

The following functions will initiate a response for the corresponding request in t. Depending on the state of the current connection, and the header values of the response, this may cause the connection to close or to persist for reuse by the client.

See RFC72306.3 for more details.

val respond_with_string : t -> Response.t -> string -> unit

val respond_with_bigstring : t -> Response.t -> Bigstringaf.t -> unit

val respond_with_streaming : ?flush_headers_immediately:bool -> t -> Response.t -> [ `write ] Body.t

Exception Handling

val report_exn : t -> exn -> unit

val try_with : t -> (unit -> unit) -> (unit, exn) Result.result