Module Httpaf.Version

Protocol Version

HTTP uses a "<major>.<minor>" numbering scheme to indicate versions of the protocol. The protocol version as a whole indicates the sender's conformance with the set of requirements laid out in that version's corresponding specification of HTTP.

See RFC72302.6 for more details.

type t = {

major : int;
The major protocol number.

minor : int;
The minor protocol number.

}

val compare : t -> t -> int

val to_string : t -> string

val of_string : string -> t

val pp_hum : Stdlib.Format.formatter -> t -> unit