Module Httpaf

Http/af is a high-performance, memory-efficient, and scalable web server for OCaml. It implements the HTTP 1.1 specification with respect to parsing, serialization, and connection pipelining. For compatibility, http/af respects the imperatives of the Server_connection header when handling HTTP 1.0 connections.

To use this library effectively, the user must be familiar with the HTTP 1.1 specification, and the basic principles of memory management and vectorized IO.

Basic HTTP Types

module Version : sig ... end

Protocol Version

module Method : sig ... end

Request Method

module Status : sig ... end

Response Status Codes

module Headers : sig ... end

Header Fields

Message Body

module Body : sig ... end

Message Types

module Request : sig ... end

Request

module Response : sig ... end

Response

module IOVec : sig ... end

IOVec

module Reqd : sig ... end

module Config : sig ... end

Server Connection

module Server_connection : sig ... end

Client Connection

module Client_connection : sig ... end