Module Httpaf.Request

Request

A client-initiated HTTP message.

type t = {

meth : Method.t;

target : string;

version : Version.t;

headers : Headers.t;

}

val create : ?⁠version:Version.t -> ?⁠headers:Headers.t -> Method.t -> string -> t

val body_length : t -> [ `Fixed of Stdlib.Int64.t | `Chunked | `Error of [ `Bad_request ] ]

body_length t is the length of the message body accompanying t. It is an error to generate a request with a close-delimited message body.

See RFC72303.3.3 for more details.

val persistent_connection : ?⁠proxy:bool -> t -> bool

persistent_connection ?proxy t indicates whether the connection for t can be reused for multiple requests and responses. If the calling code is acting as a proxy, it should pass ~proxy:true.

See RFC72306.3 for more details.

val pp_hum : Stdlib.Format.formatter -> t -> unit